requirements.txt
flask==3.0.3
python-dotenv==1.0.1

langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-pinecone==0.1.2

pinecone-client==3.2.2

sentence-transformers==2.7.0
huggingface-hub==0.23.4
torch>=2.2.0

pypdf==4.2.0

groq==0.9.0
langchain-groq==0.1.9

SQLAlchemy>=2.0.0
httpx==0.27.0